Reverted to case edit component
