+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/case-edit-confirm/case-edit-confirm.html
@@ -18,4 +18,6 @@
     <ccd-markdown [content]="confirmation.getBody()"></ccd-markdown>
   </div>
   <button type="submit" class="button" data-ng-click="submit()">{{triggerText}}</button>
+
+
 </form>
