+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/case-edit/case-edit.model.ts
@@ -34,13 +34,13 @@ export interface CaseEditGetNextPage
   extends Pick<CaseEditModel, 'wizard' | 'currentPageId'| 'eventTrigger' | 'form'> {}
 
 export interface CaseEditSubmitForm
-  extends Pick<CaseEditModel, 'eventTrigger' | 'form'| 'caseEditState' | 'caseDetails' | 'submit'>{}
+  extends Pick<CaseEditModel, 'eventTrigger' | 'form' | 'caseDetails' | 'submit'>{}
 
 export interface CaseEditGenerateCaseEventData
-  extends Pick<CaseEditModel, 'eventTrigger' | 'form'| 'caseEditState'>{}
+  extends Pick<CaseEditModel, 'eventTrigger' | 'form'>{}
 
 export interface CaseEditCaseSubmit
-  extends Pick<CaseEditModel, 'form' | 'caseEventData'| 'caseEditState' | 'submit'> {}
+  extends Pick<CaseEditModel, 'form' | 'caseEventData' | 'submit'> {}
 
 export interface CaseEditonEventCanBeCompleted
-  extends Pick<CaseEditModel, 'eventCanBeCompleted' | 'eventTrigger'| 'caseEditState' | 'caseDetails' | 'form' | 'submit'>{}
+  extends Pick<CaseEditModel, 'eventCanBeCompleted' | 'eventTrigger' | 'caseDetails' | 'form' | 'submit'>{}
